Created Price_Check_Worker and IPrice_Check_Worker
