+++ Source/Price_Check_Worker/IPrice_Check_Worker.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+class IPrice_Check_Worker(ABC):
+
+    def __init__(self):
+        None
+
+    @abstractmethod
+    def checkIsOnSale(self, symbol: str, priceData: dict) -> None:
+        pass
